rename postRoutes to GETpostRoutes
